style: Resize app_icons, Rename folders, Remove unused favicons
