--- frontend/src/components/Box.tsx
@@ -1,17 +1,0 @@
-import { useState } from "react";
-
-export default function Box({ children, smallIcon }) {
-  const onMobile = window.innerWidth < 600;
-  const [open, setOpen] = useState<boolean>(!onMobile);
-  return (
-    <>
-      <div className={`closed-box ${!open && "showing"}`} onClick={() => setOpen(true)}>
-        {smallIcon}
-      </div>
-      <div className={`open-box ${open && "showing"}`}>
-        {onMobile && <div onClick={() => setOpen(false)}>x</div>}
-        {children}
-      </div>
-    </>
-  );
-}
