--- frontend/src/components/BoxContainer.tsx
@@ -1,3 +1,0 @@
-export default function BoxContainer({ children }) {
-  return <div className="fullscreen over-boxes">{children}</div>;
-}
