--- frontend/src/components/DesktopApps/Screens/HorizontalBar.tsx
@@ -1,3 +1,0 @@
-export default function HorizontalBar() {
-  return <div style={{ width: "100%", flex: "3%", backgroundColor: "black" }}></div>;
-}
