--- frontend/src/components/DesktopApps/Screens/VerticalBar.tsx
@@ -1,3 +1,0 @@
-export default function VerticalBar() {
-  return <div style={{ height: "100%", flex: "2%", backgroundColor: "black" }}></div>;
-}
