+++ frontend/src/components/desktop_apps/screens/BitMinerAppScreen.tsx
