+++ frontend/src/components/desktop_apps/screens/HorizontalBar.tsx
@@ -1,0 +1,3 @@
+export default function HorizontalBar() {
+  return <div style={{ width: "100%", flex: "3%", backgroundColor: "black" }}></div>;
+}
