+++ frontend/src/components/desktop_apps/screens/UpgradeAppScreen.tsx
