+++ frontend/src/components/desktop_apps/screens/VerticalBar.tsx
@@ -1,0 +1,3 @@
+export default function VerticalBar() {
+  return <div style={{ height: "100%", flex: "2%", backgroundColor: "black" }}></div>;
+}
